% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expected_rank}
\alias{expected_rank}
\title{Expected rank}
\usage{
expected_rank(mat)
}
\arguments{
\item{mat}{A matrix of samples}
}
\value{
A vector of size ncol(mat) giving expected rank for each column
}
\description{
Calculate expected ranks according to posterior draws.
}
