% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findfirst}
\alias{findfirst}
\title{Find first element of vector satisfying condition}
\usage{
findfirst(x, v = NA)
}
\arguments{
\item{x}{Logical vector}

\item{v}{Value of NA}
}
\value{
First element satisfying condition
}
\description{
Find first element of vector satisfying condition
}
