% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mvn_entropy}
\alias{mvn_entropy}
\title{Calculate the differential entropy of multivariate normal}
\usage{
mvn_entropy(S)
}
\arguments{
\item{S}{A square matrix}
}
\value{
The log-determinant of S
}
\description{
Calculate the differential entropy of multivariate normal
}
