% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pairwise_diff}
\alias{pairwise_diff}
\title{Pairwise difference matrix}
\usage{
pairwise_diff(mat)
}
\arguments{
\item{mat}{A matrix of samples}
}
\value{
A matrix of K(K-1)/2 pairwise differences
}
\description{
A function to generate pairwise difference matrix for unique combinations.
The function only generates the lower triangular differences, that is K(K-1)/2
where K is the number of columns.
}
