% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pairwise_diff_all}
\alias{pairwise_diff_all}
\title{Pairwise difference matrix}
\usage{
pairwise_diff_all(mat, ...)
}
\arguments{
\item{mat}{A matrix of samples}
}
\value{
A matrix of K(K-1) pairwise differences in both directions.
}
\description{
A function to generate pairwise difference matrix for all combinations.
The function generates all K(K-1) differences (excludes the diagonal)
where K is the number of columns.
}
