% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pairwise_superiority_all}
\alias{pairwise_superiority_all}
\title{Pairwise superiority for both directions}
\usage{
pairwise_superiority_all(mat, delta = 0, ...)
}
\arguments{
\item{mat}{A matrix of samples}

\item{delta}{Reference value of interest (negative for noninferiority)}
}
\value{
A vector of K(K-1) pairwise probabiliy of superiority
}
\description{
Pairwise superiority for all K*(K-1) comparisons rather than
the reduced K(K-1)/2 one-way comparisons.
}
