% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prob_all_equivalent}
\alias{prob_all_equivalent}
\title{Probability equivalent}
\usage{
prob_all_equivalent(mat, delta)
}
\arguments{
\item{mat}{A matrix of samples}

\item{delta}{Equivalence bound}
}
\description{
Probability all arms in mat are equivalent within delta
}
