% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prob_all_superior}
\alias{prob_all_superior}
\title{Probability superior}
\usage{
prob_all_superior(a, b, delta = 0)
}
\arguments{
\item{a}{A matrix}

\item{b}{A vector}

\item{delta}{The minimal effect bound}
}
\value{
A numeric probability
}
\description{
Probability that all in a are superior b
}
\examples{
# All superior
prob_all_superior(matrix(rnorm(1e4, 1:10), 1000, 10, byrow = T), rnorm(1e4, 0), 0.5)
# All non-inferior
prob_all_superior(matrix(rnorm(1e4, 1:10), 1000, 10, byrow = T), rnorm(1e4, 0), -0.5)
}
