% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prob_each_superior_all}
\alias{prob_each_superior_all}
\title{Probability each superior to all}
\usage{
prob_each_superior_all(mat, delta = 0)
}
\arguments{
\item{mat}{A matrix of samples}

\item{delta}{The minimal effect bound}
}
\value{
A vector of probabilities that each dimension is maximum
}
\description{
Probability that each column in mat is superior to all other columns.
}
\examples{
# Superiority
prob_each_superior_all(matrix(rnorm(1e4, 1:10), 1000, 10, byrow = T), 0.5)
# Non-inferiority
prob_each_superior_all(matrix(rnorm(1e4, 1:10), 1000, 10, byrow = T), -0.5)
}
