% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prob_h_indiff}
\alias{prob_h_indiff}
\title{Probability that the h best populations ranked according to ord
are the best h.}
\usage{
prob_h_indiff(mat, r = order(expected_rank(mat)), delta)
}
\arguments{
\item{mat}{A matrix of samples}

\item{r}{A ranking vector giving the order from smallest to largest}

\item{delta}{An indifference zone value}
}
\value{
A vector giving the probability of indifference (equivalence) 
between the h ranked means
}
\description{
Probability that the h best populations ranked according to ord
are the best h.
}
\examples{
# Expect high probability that two best identified
# Others should be 50/50
D <- mvnfast::rmvn(4e4, c(1,1,2,2), diag(0.1,4))
rbind(prob_h_best(D), prob_h_indiff(D, delta = 1))
}
