% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prob_max}
\alias{prob_max}
\title{Probability that each column in mat is the maximum of all columns}
\usage{
prob_max(mat)
}
\arguments{
\item{mat}{A matrix of samples}
}
\value{
A vector of probabilities that each dimension is maximum
}
\description{
Probability that each column in mat is the maximum of all columns
}
\examples{
prob_max(matrix(rnorm(1e4, 1:10), 1000, 10, byrow = T))
}
