% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prob_superior}
\alias{prob_superior}
\title{Probability that items in a are superior to b}
\usage{
prob_superior(a, b, delta = 0)
}
\arguments{
\item{a}{A matrix or vector of items to compare to b}

\item{b}{A vector or matrix to which a is compared}

\item{delta}{The minimal effect bound}
}
\value{
A vector of probabilities
}
\description{
Probability that items in a are superior to b
}
\examples{
prob_superior(mvnfast::rmvn(1e4, 1:3, diag(1, 3)), rnorm(1e4), 1)
prob_superior(rnorm(1e4), mvnfast::rmvn(1e4, 1:3, diag(1, 3)), 1)
}
