% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prob_superior_all}
\alias{prob_superior_all}
\title{Probability superior to all}
\usage{
prob_superior_all(a, b, delta = 0)
}
\arguments{
\item{a}{A vector}

\item{b}{A matrix}

\item{delta}{The minimal effect bound}
}
\value{
A numeric probability
}
\description{
Probability that a is superior to all items in b
}
