% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_noninferiority_alt.R
\name{run_a_noninf_trial_alt}
\alias{run_a_noninf_trial_alt}
\title{Run an alternative non-inferiority trial}
\usage{
run_a_noninf_trial_alt(id, mu, delta_sup = 0.2, delta_ctr = delta_sup,
  delta_non = delta_sup, kappa_lo_0 = 0.01, kappa_lo_1 = 0.01,
  kappa_hi_0 = 0.9, kappa_hi_1 = 0.8, kappa_no_0 = 0.6,
  kappa_no_1 = 0.6, kappa_ctrl = 0.95, brar = TRUE,
  allocate_inactive = FALSE, return_all = FALSE, ctrl_alloc = 1/13)
}
\arguments{
\item{id}{The trial ID}

\item{mu}{The true mean, must be length 13}

\item{kappa_lo_0}{The starting threshold to deactivate poor arms}

\item{kappa_lo_1}{The final threshold to deactivate poor arms}

\item{kappa_hi_0}{The starting threshold for superiority}

\item{kappa_hi_1}{The final threshold for superiority}

\item{kappa_no_0}{The starting threshold for non-inferiority}

\item{kappa_no_1}{The final threshold for non-inferiority}

\item{kappa_ctrl}{The control comparison threshold}

\item{brar}{Use Response Adaptive Randomisation?}

\item{allocate_inactive}{Continue to allocate subjects to arms meeting the inactive threshold?}

\item{return_all}{Return all value, or only those from final analysis?}

\item{delta}{The superiority margin}
}
\value{
A list of trial quantities
}
\description{
Run a single trial using non-inferiority stopping.
Stopping for noninferiority requires that:
}
