% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trial_nointerims.R
\name{run_a_noninf_trial_nointerims}
\alias{run_a_noninf_trial_nointerims}
\title{Run a non-inferiority trial}
\usage{
run_a_noninf_trial_nointerims(
  id,
  mu,
  delta,
  kappa_lo = 0.01,
  kappa_hi = 0.75,
  kappa_no = 0.5,
  kappa_ctrl = 0.95,
  ctrl_alloc = 1/13,
  ind_comp_ctrl = FALSE
)
}
\arguments{
\item{id}{The trial ID}

\item{mu}{The true mean, must be length 13}

\item{delta}{The superiority margin}

\item{kappa_lo}{The starting threshold to deactivate poor arms}

\item{kappa_hi}{The starting threshold for superiority}

\item{kappa_no}{The starting threshold for non-inferiority}

\item{kappa_ctrl}{The starting threshold for non-inferiority}

\item{ind_comp_ctrl}{Should deactivation be based on P(max trt arms) or P(max all arms)}
}
\value{
A list of trial quantities
}
\description{
Run a single trial with no interims.
}
