% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_nonsuperiority.R
\name{run_a_nonsup_trial}
\alias{run_a_nonsup_trial}
\title{Run a non-superiority trial}
\usage{
run_a_nonsup_trial(
  id,
  mu,
  delta,
  kappa_lo_0 = 0.01,
  kappa_lo_1 = 0.05,
  kappa_hi_0 = 0.9,
  kappa_hi_1 = 0.85,
  brar = FALSE,
  allocate_inactive = FALSE,
  return_all = FALSE
)
}
\arguments{
\item{id}{The trial ID}

\item{mu}{The true mean, must be length 13}

\item{delta}{The superiority margin}

\item{kappa_lo_0}{The starting threshold to deactivate poor arms}

\item{kappa_lo_1}{The final threshold to deactivate poor arms}

\item{kappa_hi_0}{The starting threshold for superiority}

\item{kappa_hi_1}{The final threshold for superiority}

\item{brar}{Use Response Adaptive Randomisation?}

\item{allocate_inactive}{Continue to allocate subjects to arms meeting the inactive threshold?}

\item{return_all}{Return all value, or only those from final analysis?}
}
\value{
A list of trial quantities
}
\description{
Run a single trial using non-superiority stopping.
The idea is as follows:
}
