% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_automatic.R
\name{run_automatic_trial}
\alias{run_automatic_trial}
\title{Run automatic trial simulation}
\usage{
run_automatic_trial(
  mu,
  Sigma0 = diag(c(2^2, rep(1, 12))),
  n_seq = seq(500, 10000, 500),
  delta_sup = log(1.1),
  thres_sup = 0.85,
  thres_inf = 1 - thres_sup,
  thres_fut = 0.95,
  thres_nin = 0.9,
  thres_eff = 0.99,
  thres_equ = 0.9,
  thres_ina = 0.99,
  rar_scale = 0.5,
  use_mwu = TRUE,
  mc_draws = 10000,
  ct_alloc = 0.25
)
}
\arguments{
\item{mu}{The true response rate in each arm}

\item{n_seq}{The interim sequence}

\item{delta_sup}{The reference value for meaningful difference}

\item{thres_sup}{Threshold for superiority}

\item{thres_inf}{Threshold for inferiority}

\item{thres_fut}{Threshold for futility}

\item{thres_nin}{Threshold for non-inferiority}

\item{thres_eff}{Threshold for effectiveness}

\item{thres_equ}{Threshold for equality}

\item{thres_ina}{Threshold for inadequacy}

\item{use_mwu}{Use mass-weighted-urn randomisation}

\item{mc_draws}{Number of Monte Carlo draws for posterior quantity calculation}
}
\description{
Run automatic trial simulation
}
