% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trial_noninterims_indifference.R
\name{run_indifference_trial_nointerims}
\alias{run_indifference_trial_nointerims}
\title{Run an indifference trial with no interims}
\usage{
run_indifference_trial_nointerims(
  id,
  mu,
  delta_indiff = 0.25,
  delta_ctrl = delta_indiff,
  kappa_sup = 0.95,
  kappa_ind = 0.75,
  kappa_ctr = 0.95,
  ctrl_alloc = 1/13
)
}
\description{
An indifference stopping...
}
