% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{thres_seq}
\alias{thres_seq}
\title{Threshold sequence}
\usage{
thres_seq(a0, a1, r, t)
}
\arguments{
\item{a0}{The starting threshold}

\item{a1}{The ending threshold}

\item{r}{The change transformation}

\item{t}{The sequence length from start to end}
}
\value{
A vector of length t + 1 giving the threshold sequence
}
\description{
Threshold sequence
}
