% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_automatic.R
\name{tibble_ctr_quantities}
\alias{tibble_ctr_quantities}
\title{Group list of trial outcomes into a tibble}
\usage{
tibble_ctr_quantities(dat, final = TRUE, ...)
}
\arguments{
\item{dat}{The results of `run_automatic_trial` as a list}

\item{final}{Return data from final analysis or interims}

\item{...}{Other arguments to `mclapply`}
}
\description{
Group list of trial outcomes into a tibble
}
