% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_automatic.R
\name{tibble_dec1_quantities}
\alias{tibble_dec1_quantities}
\title{Group list of decision outcomes into a tibble}
\usage{
tibble_dec1_quantities(dat, ...)
}
\arguments{
\item{dat}{The results of `ascot_trial2` as a list}

\item{...}{Other arguments to `mclapply`}
}
\description{
Group list of decision outcomes into a tibble
}
