% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_automatic.R
\name{tibble_result_quantities}
\alias{tibble_result_quantities}
\title{Group list of trial outcomes into a tibble}
\usage{
tibble_result_quantities(dat, ...)
}
\arguments{
\item{dat}{The results of `ascot_trial2` as a list}

\item{...}{Other arguments to `mclapply`}
}
\description{
Group list of trial outcomes into a tibble
}
