% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_automatic.R
\name{run_automatic_trial}
\alias{run_automatic_trial}
\title{Run automatic trial simulation}
\usage{
run_automatic_trial(
  mu,
  Xdes = 1,
  mu0 = c(qlogis(0.8), rep(0, 12)),
  Sigma0 = diag(c(2^2, rep(1, 12))),
  n_seq = seq(500, 10000, 500),
  delta_sup = log(1.1),
  thres_sup = function(t) 0.8,
  thres_inf = function(t) 1 - thres_sup(t),
  thres_fut = 0.95,
  thres_nin = 0.9,
  thres_eff = function(t) 0.99,
  thres_hrm = function(t) 0.01,
  thres_equ = 0.9,
  thres_ina = function(t) 0.99,
  thres_ctr = function(t) 0.99,
  use_mwu = TRUE,
  mc_draws = 10000,
  rar_scale = 0.5,
  rar_min = 0.05/12,
  rar_rule = 1,
  drop_rule = 1,
  stop_rule = 1,
  ct_alloc = sqrt(12)/(12 + sqrt(12)),
  drop_perm = TRUE,
  enable_stopping = FALSE,
  return_all = TRUE,
  gen_prior = FALSE,
  use_info_n = TRUE
)
}
\arguments{
\item{mu}{The true response rate in each arm}

\item{n_seq}{The interim sequence}

\item{delta_sup}{The reference value for meaningful difference}

\item{thres_sup}{Threshold for superiority}

\item{thres_inf}{Threshold for inferiority}

\item{thres_fut}{Threshold for futility}

\item{thres_nin}{Threshold for non-inferiority}

\item{thres_eff}{Threshold for effectiveness}

\item{thres_equ}{Threshold for equality}

\item{thres_ina}{Threshold for inadequacy}

\item{use_mwu}{Use mass-weighted-urn randomisation}

\item{mc_draws}{Number of Monte Carlo draws for posterior quantity calculation}

\item{rar_scale}{Scaling factor for RAR}

\item{rar_rule}{Which RAR rule to use}

\item{drop_rule}{Which rule to use for dropping control}

\item{ct_alloc}{What fixed allocation to control}

\item{drop_perm}{Permanently drop interventions?}

\item{enable_stopping}{Apply stopping rule or let trial run its course to max SS?}
}
\description{
Run automatic trial simulation
}
